提现规则

+++ city-base/src/main/java/com/dream/city/base/model/entity/PlayerEarning.java
@@ -1,8 +1,12 @@
 package com.dream.city.base.model.entity;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.Date;
 
+/**
+ * 玩家提现规则
+ */
 public class PlayerEarning implements Serializable {
     /**  */
     private Integer earnId;
@@ -11,10 +15,10 @@ public class PlayerEarning implements Serializable {
     private String earnPlayerId;
 
     /** 最大提取额度 */
-    private Double earnMax;
+    private BigDecimal earnMax;
 
     /** 税金 */
-    private Double earnTax;
+    private BigDecimal earnTax;
 
     private Date createTime;
 
@@ -34,19 +38,19 @@ public class PlayerEarning implements Serializable {
         this.earnPlayerId = earnPlayerId == null ? null : earnPlayerId.trim();
     }
 
-    public Double getEarnMax() {
+    public BigDecimal getEarnMax() {
         return earnMax;
     }
 
-    public void setEarnMax(Double earnMax) {
+    public void setEarnMax(BigDecimal earnMax) {
         this.earnMax = earnMax;
     }
 
-    public Double getEarnTax() {
+    public BigDecimal getEarnTax() {
         return earnTax;
     }
 
-    public void setEarnTax(Double earnTax) {
+    public void setEarnTax(BigDecimal earnTax) {
         this.earnTax = earnTax;
     }
